Create a new SettingsManager. Similar to Logger, it is a singleton. Implement it in ObsidianIcalPlugin so that it can be used anywhere. Change everywhere so that we are using the new getSetting and setSetting functions.
Now SettingsManager acts as a cache layer between the application and the plugin saveData and loadData functions.

Fixes a bug where settings were not loading properly when new settings were added after a version upgrade.

Fixes a bug where changing a value of a setting was not reflected through the code.